requirements.txt
openai
google-generativeai
rich
python-dotenv
